Refactor object_set to tileset
