+++ foundry/smb3parse/objects/tileset.py
@@ -31,10 +31,15 @@ MAX_OBJECT_SET = 0x0F
 OBJECT_GROUP_SIZE = 16
 
 
-def assert_valid_object_set_number(object_set_number: int):
-    if not is_valid_object_set_number(object_set_number):
-        raise ValueError(f"Object set number {object_set_number} is invalid.")
+class TilesetError(ValueError):
+    def __init__(self, tileset_index: int) -> None:
+        super().__init__(f"Tileset is at invalid index: {tileset_index}.")
 
 
-def is_valid_object_set_number(object_set_number: int):
-    return object_set_number in range(MIN_OBJECT_SET, MAX_OBJECT_SET + 1)
+def ensure_tileset(tileset_index: int) -> None:
+    if not is_tileset_index(tileset_index):
+        raise TilesetError(tileset_index)
+
+
+def is_tileset_index(tileset_index: int):
+    return tileset_index in range(MIN_OBJECT_SET, MAX_OBJECT_SET + 1)
