+++ tests/game/level/test_level.py
@@ -5,7 +5,7 @@ from foundry.game.gfx.objects.EnemyItem import EnemyObject
 from foundry.game.gfx.objects.Jump import Jump
 from foundry.game.gfx.objects.LevelObject import LevelObject
 from foundry.game.level.Level import LEVEL_DEFAULT_HEIGHT, Level
-from foundry.smb3parse.objects.object_set import PLAINS_OBJECT_SET
+from foundry.smb3parse.objects.tileset import PLAINS_OBJECT_SET
 from tests.conftest import level_1_1_enemy_address, level_1_1_object_address
