+++ tests/gui/test_object_dropdown.py
@@ -1,4 +1,4 @@
-from foundry.smb3parse.objects.object_set import HILLY_OBJECT_SET
+from foundry.smb3parse.objects.tileset import HILLY_OBJECT_SET
 from tests.conftest import level_1_2_enemy_address, level_1_2_object_address
